return conitionally rendered business registration message to the way it initially worked. Currently, business registration form reappears if the user clicks on the register business link again.
